requirements.txt
langchain
langchain-openai
langchain_community
faiss-cpu
langchain_huggingface
langchain_core
langchain_google_firestore
firebase_admin
sentence-transformers
flask
google-auth